requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
anthropic>=0.18.0
slowapi>=0.1.9
prometheus-client>=0.19.0
requests>=2.31.0
